+++ P4/arqo4.h
@@ -8,12 +8,12 @@
 #define N 1000ull
 #define M 1000000ull
 
-float ** generateMatrix(int);
-float ** generateEmptyMatrix(int);
+float ** generateMatrix(unsigned long long);
+float ** generateEmptyMatrix(unsigned long long);
 void freeMatrix(float **);
-float * generateVector(int);
-float * generateEmptyVector(int);
-int * generateEmptyIntVector(int);
+float * generateVector(unsigned long long);
+float * generateEmptyVector(unsigned long long);
+int * generateEmptyIntVector(unsigned long long);
 void freeVector(void *);
 
 #endif /* _ARQO_P4_H_ */
